+++ packages/hardhat/test/fixture-projects/hardhat-project/contracts/EdgeCases.sol
@@ -3,11 +3,6 @@ pragma solidity 0.7.3;
 pragma experimental ABIEncoderV2;
 
 contract EdgeCases {
-  // fix for: https://github.com/ethereum-ts/TypeChain/issues/389
-  constructor (
-    uint256 test,
-    bytes memory
-  ) {
-
-  }
+  // fix for: https://github.com/gnosisguild/TypeChain/issues/389
+  constructor(uint256 test, bytes memory) {}
 }