+++ packages/target-ethers-v6/test/typescript-version.test.ts
@@ -32,11 +32,13 @@ describe('Ethers target constructor', () => {
         allFiles: ['woop'],
         target: 'ethers-v6',
       })
-    }).toThrow('@typechain/ethers-v6 1.2.3 needs TypeScript version 4.3 or newer.')
+    }).toThrow('@gnosis-guild/typechain-ethers-v6 1.2.3 needs TypeScript version 4.3 or newer.')
 
     const message = consoleErrorMock.calls[0].args[0]
 
-    expect(message).toEqual(expect.stringMatching('@typechain/ethers-v6 1.2.3 needs TypeScript version 4.3 or newer.'))
+    expect(message).toEqual(
+      expect.stringMatching('@gnosis-guild/typechain-ethers-v6 1.2.3 needs TypeScript version 4.3 or newer.'),
+    )
     expect(message).toEqual(
       expect.stringMatching('Generated code will cause syntax errors in older TypeScript versions.'),
     )