+++ packages/target-starknet/src/index.ts
@@ -1,3 +1,0 @@
-import { StarknetTarget } from './StarknetTarget'
-
-export default StarknetTarget